v1.3 Build - Release Candidate
*This is a potential Release Candidate, unless another is uploaded.

Fixed credits and made hints easier to read in the moment.
Tweaked some block boxes and geometry in Overlook.
Trying to get the fucking ambience to change in Overlook is a pain in my ass.
Added block boxes + statues to block the path after the "cavein" event, as to prevent backtracking.